Create an actual package to organize the pybin source
