+++ src/pybin/dive/build.py
@@ -1,4 +1,4 @@
-from ..buildlib import build_wheels
+from pybin.buildlib import build_wheels
 
 NAME = 'dive'
 UPSTREAM_REPO = 'https://github.com/wagoodman/dive'
