+++ src/pybin/hadolint/build.py
@@ -1,4 +1,4 @@
-from ..buildlib import build_wheels
+from pybin.buildlib import build_wheels
 
 NAME = 'hadolint'
 UPSTREAM_REPO = 'https://github.com/hadolint/hadolint'
