+++ src/pybin/scc/build.py
@@ -1,10 +1,10 @@
-from ..buildlib import build_wheels
+from pybin.buildlib import build_wheels
 
-NAME = 'lazydocker'
-UPSTREAM_REPO = 'https://github.com/jesseduffield/lazydocker'
-VERSION = '0.23.1'
-PYPI_VERSION = '0.23.1'
-LICENSE = "MIT"
+NAME = 'scc'
+UPSTREAM_REPO = 'https://github.com/boyter/scc'
+VERSION = '3.2.0'
+PYPI_VERSION = '3.2.0'
+LICENSE = "The Unlicense (Unlicense)"
 
 TARGET_TAG = {
     'Darwin_arm64': 'macosx_11_0_arm64',
@@ -12,7 +12,7 @@ TARGET_TAG = {
     'Linux_arm64': 'manylinux_2_17_aarch64.manylinux2014_aarch64.musllinux_1_1_aarch64',
     'Linux_x86_64': 'manylinux_2_12_x86_64.manylinux2010_x86_64.musllinux_1_1_x86_64',
 }
-URL_TAG = {f"{UPSTREAM_REPO}/releases/download/v{VERSION}/{NAME}_{VERSION}_{target}.tar.gz": tag for target, tag in TARGET_TAG.items()}
+URL_TAG = {f"{UPSTREAM_REPO}/releases/download/v{VERSION}/{NAME}_{target}.tar.gz": tag for target, tag in TARGET_TAG.items()}
 
 
 if __name__ == "__main__":
